old/project_hangarr.__backup__/projects/2026-04_peak-hub/tools/requirements.txt
playwright
python-dotenv
google-api-python-client>=2.130.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
